# Reduce allocations in common paths