--- a/src/MySqlConnector/Core/ServerSession.cs
+++ b/src/MySqlConnector/Core/ServerSession.cs
@@ -319,13 +319,12 @@ namespace MySqlConnector.Core
 				{
 					m_logArguments[1] = ServerVersion.OriginalString;
 					Log.Debug("{0} ServerVersion {1} supports reset connection; sending reset connection request", m_logArguments);
-					await SendAsync(ResetConnectionPayload.Create(), ioBehavior, cancellationToken).ConfigureAwait(false);
+					await SendAsync(ResetConnectionPayload.Instance, ioBehavior, cancellationToken).ConfigureAwait(false);
 					var payload = await ReceiveReplyAsync(ioBehavior, cancellationToken).ConfigureAwait(false);
 					OkPayload.Create(payload);
 
 					// the "reset connection" packet also resets the connection charset, so we need to change that back to our default
-					payload = QueryPayload.Create("SET NAMES utf8mb4 COLLATE utf8mb4_bin;");
-					await SendAsync(payload, ioBehavior, cancellationToken).ConfigureAwait(false);
+					await SendAsync(s_setNamesUtf8mb4Payload, ioBehavior, cancellationToken).ConfigureAwait(false);
 					payload = await ReceiveReplyAsync(ioBehavior, cancellationToken).ConfigureAwait(false);
 					OkPayload.Create(payload);
 				}
@@ -1151,6 +1150,7 @@ namespace MySqlConnector.Core
 		static int s_lastId;
 		static byte[] s_connectionAttributes;
 		static readonly IMySqlConnectorLogger Log = MySqlConnectorLogManager.CreateLogger(nameof(ServerSession));
+		static readonly PayloadData s_setNamesUtf8mb4Payload = QueryPayload.Create("SET NAMES utf8mb4 COLLATE utf8mb4_bin;");
 
 		readonly object m_lock;
 		readonly object[] m_logArguments;