--- a/src/MySqlConnector/Protocol/Payloads/PingPayload.cs
+++ b/src/MySqlConnector/Protocol/Payloads/PingPayload.cs
@@ -2,6 +2,6 @@ namespace MySqlConnector.Protocol.Payloads
 {
 	internal sealed class PingPayload
 	{
-		public static PayloadData Create() => new PayloadData(new[] { (byte) CommandKind.Ping });
+		public static PayloadData Instance { get; } = new PayloadData(new[] { (byte) CommandKind.Ping });
 	}
 }