--- a/src/MySqlConnector/Protocol/Payloads/QuitPayload.cs
+++ b/src/MySqlConnector/Protocol/Payloads/QuitPayload.cs
@@ -2,6 +2,6 @@ namespace MySqlConnector.Protocol.Payloads
 {
 	internal sealed class QuitPayload
 	{
-		public static PayloadData Create() => new PayloadData(new[] { (byte) CommandKind.Quit });
+		public static PayloadData Instance { get; } = new PayloadData(new[] { (byte) CommandKind.Quit });
 	}
 }