```
BenchmarkDotNet=v0.10.12, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.248)
Intel Xeon W-2155 CPU 3.30GHz, 1 CPU, 20 logical cores and 10 physical cores
Frequency=3234375 Hz, Resolution=309.1787 ns, Timer=TSC
  [Host]    : .NET Framework 4.7 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.2633.0
  net47     : .NET Framework 4.7 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.2633.0
  netcore20 : .NET Core 2.0.5 (Framework 4.6.26020.03), 64bit RyuJIT
```

Connection String: `server=127.0.0.1;user id=mysqltest;password='test;key=\"val';database=hello_world;ssl mode=none;Default Command Timeout=0;Connection Timeout=0;Connection Reset=true`

## Before

Job | Runtime |     Toolchain |     Mean |    Error |   StdDev |    StdErr |      Min |       Q1 |   Median |       Q3 |      Max |    Op/s |  Gen 0 | Allocated |
---------- |-------- |-------------- |---------:|---------:|---------:|----------:|---------:|---------:|---------:|---------:|---------:|--------:|-------:|----------:|
net47 |     Clr |   CsProjnet47 | 290.6 us | 2.295 us | 2.147 us | 0.5543 us | 286.2 us | 289.6 us | 290.3 us | 292.1 us | 294.2 us | 3,441.5 | 2.4414 |  16.18 KB |
netcore20 |    Core | .NET Core 2.0 | 161.2 us | 1.282 us | 1.200 us | 0.3097 us | 158.7 us | 160.1 us | 161.6 us | 161.8 us | 162.9 us | 6,202.7 | 1.7090 |   4.52 KB |

## After

Job | Runtime |     Toolchain |     Mean |    Error |   StdDev |    StdErr |      Min |       Q1 |   Median |       Q3 |      Max |    Op/s |  Gen 0 | Allocated |
---------- |-------- |-------------- |---------:|---------:|---------:|----------:|---------:|---------:|---------:|---------:|---------:|--------:|-------:|----------:|
net47 |     Clr |   CsProjnet47 | 289.0 us | 2.312 us | 2.162 us | 0.5583 us | 286.2 us | 287.3 us | 288.7 us | 290.2 us | 293.4 us | 3,459.8 | 2.4414 |  15.06 KB |
netcore20 |    Core | .NET Core 2.0 | 163.1 us | 2.095 us | 1.959 us | 0.5059 us | 158.8 us | 162.3 us | 162.9 us | 164.7 us | 166.8 us | 6,130.2 | 1.7090 |   3.76 KB |

Memory usage is 93% of previous (.NET 4.7), 83% of previous (.NET Core).